requirements.txt
aiogoogletrans==3.3.2
aiohttp>=3.7.4
arrow==1.1.0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
chardet==4.0.0
click==8.0.0
colorama==0.4.4
dnspython==2.1.0
geographiclib==1.50
geopy==2.1.0
gql==3.0.0a5
graphql-core==3.1.5
html5lib==1.1
idna==3.1
joblib==1.0.1
markovify==0.9.0
motor==2.4.0
multidict==5.1.0
nltk==3.6.2
numpy==1.20.3
Pillow==8.3.2
pymongo==3.11.4
python-dateutil==2.8.1
PyYAML==5.4.1
regex==2021.4.4
six==1.16.0
soupsieve==2.2.1
tqdm==4.60.0
typing-extensions==3.10.0.0
Unidecode==1.2.0
webencodings==0.5.1
websockets>=9.1
yarl==1.6.3
youtube_dl
git+https://github.com/Pycord-Development/pycord@feature/slash#egg=py-cord